requirements.txt
netcdf4
dask
numpy
pandas
pyyaml
xarray
xgboost
scikit-learn
scipy
scikit-bio
